+++ ui/pages/Project/components/Matchmaker.jsx
@@ -37,7 +37,7 @@ import {
   getMmeResultsBySubmission,
   getMmeDefaultContactEmail,
   getMatchmakerContactNotes,
-  getVariantUniqueId,
+  getVariantGeneId,
 } from '../selectors'
 import SelectSavedVariantsTable from './SelectSavedVariantsTable'
 
@@ -133,7 +133,7 @@ const SUBMISSION_EDIT_FIELDS = [
     includeSelectedRowData: true,
     parse: val => Object.values(val || {}).filter(v => v),
     format: value => (value || []).reduce(
-      (acc, variant) => ({ ...acc, [variant.variantId || getVariantUniqueId(variant)]: variant }), {},
+      (acc, variant) => ({ ...acc, [getVariantGeneId(variant)]: variant }), {},
     ),
   },
   {